Fixed- Sell Archive Deal Delete Operation Blocked

+++ src/AppBundle/Controller/Api/BookDealManagementApiController.php
@@ -749,20 +749,27 @@ class BookDealManagementApiController extends Controller
 
         $bookDeal = $bookDealRepo->findOneById($data['bookDealId']);
         if($bookDeal->getSeller()->getId()==$userId){
-            $em->remove($bookDeal);
-            $em->flush();
+            if($bookDeal->getBuyer()!==null){
+                return $this->_createJsonResponse('error', array(
+                'errorTitle' => "Cannot Delete Book Deal",
+                'errorDescription' => "This book is sold."
+                ), 400);
+            }else {
+                $em->remove($bookDeal);
+                $em->flush();
 
-            $logData = array(
-                'user'=>$this->get('security.token_storage')->getToken()->getUser()->getId(),
-                'logType'=>"Delete Book Deal",
-                'logDateTime'=>gmdate('Y-m-d H:i:s'),
-                'logDescription'=> $this->get('security.token_storage')->getToken()->getUser()->getUsername()." has deleted a book deal named ".$bookDeal->getBook()->getBookTitle(),
-                'userIpAddress'=>$this->container->get('request')->getClientIp(),
-                'logUserType'=> in_array("ROLE_ADMIN_USER",$this->get('security.token_storage')->getToken()->getUser()->getRoles())?"Admin User":"Normal User"
-            );
-            $this->_saveLog($logData);
+                $logData = array(
+                    'user' => $this->get('security.token_storage')->getToken()->getUser()->getId(),
+                    'logType' => "Delete Book Deal",
+                    'logDateTime' => gmdate('Y-m-d H:i:s'),
+                    'logDescription' => $this->get('security.token_storage')->getToken()->getUser()->getUsername() . " has deleted a book deal named " . $bookDeal->getBook()->getBookTitle(),
+                    'userIpAddress' => $this->container->get('request')->getClientIp(),
+                    'logUserType' => in_array("ROLE_ADMIN_USER", $this->get('security.token_storage')->getToken()->getUser()->getRoles()) ? "Admin User" : "Normal User"
+                );
+                $this->_saveLog($logData);
 
-            return $this->_createJsonResponse('success', array('successTitle' => "Book Deal is Deleted"), 200);
+                return $this->_createJsonResponse('success', array('successTitle' => "Book Deal is Deleted"), 200);
+            }
         }else{
             return $this->_createJsonResponse('error', array('errorTitle' => "Cannot Delete Book Deal", 'errorDescription' => "You are not owner of that book deal."), 400);
         }
